requirements.txt
asgiref==3.8.1
certifi==2024.8.30
dj-database-url==2.2.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
Pillow==11.0.0
PyJWT==2.8.0
pytz==2024.2
whitenoise==6.5.0
psycopg2==2.9.10
django-environ==0.11.2
pytest
pytest-django